add exit code

+++ main.ps1
@@ -1,13 +1,46 @@
 #!/bin/pwsh
 
 docker pull acblog/wasm:latest
+if (!$?) {
+    exit 1
+}
 $container=$(docker run -d acblog/wasm:latest)
+if (!$?) {
+    exit 1
+}
 docker stop $container
+if (!$?) {
+    exit 1
+}
 docker cp ${container}:/app $env:INPUT_DIST
+if (!$?) {
+    exit 1
+}
 docker rm $container
-cp -r $env:INPUT_APP/* $env:INPUT_DIST
-echo "* binary" > $env:INPUT_DIST/.gitattributes
-echo "" > $env:INPUT_DIST/.nojekyll
-$(cat $env:INPUT_DIST/404.html) -replace "<base href=""/"" />", "<base href=""$env:INPUT_DOMAIN"" />" > $env:INPUT_DIST/404.html
-$(cat $env:INPUT_DIST/404.html) -replace "sparedirectEncode(l, 0)", "sparedirectEncode(l, $env:INPUT_SEGMENTCOUNT)" > $env:INPUT_DIST/404.html
-$(cat $env:INPUT_DIST/index.html) -replace "<base href=""/"" />", "<base href=""$env:INPUT_DOMAIN"" />" > $env:INPUT_DIST/index.html
+if (!$?) {
+    exit 1
+}
+Copy-Item -Recurse $env:INPUT_APP/* $env:INPUT_DIST
+if (!$?) {
+    exit 1
+}
+Write-Output "* binary" > $env:INPUT_DIST/.gitattributes
+if (!$?) {
+    exit 1
+}
+Write-Output "" > $env:INPUT_DIST/.nojekyll
+if (!$?) {
+    exit 1
+}
+$(Get-Content $env:INPUT_DIST/404.html) -Replace "<base href=""/"" />", "<base href=""$env:INPUT_DOMAIN"" />" > $env:INPUT_DIST/404.html
+if (!$?) {
+    exit 1
+}
+$(Get-Conten $env:INPUT_DIST/404.html) -Replace "sparedirectEncode(l, 0)", "sparedirectEncode(l, $env:INPUT_SEGMENTCOUNT)" > $env:INPUT_DIST/404.html
+if (!$?) {
+    exit 1
+}
+$(Get-Conten $env:INPUT_DIST/index.html) -Replace "<base href=""/"" />", "<base href=""$env:INPUT_DOMAIN"" />" > $env:INPUT_DIST/index.html
+if (!$?) {
+    exit 1
+}
